docs: reservation docs 수정
