--- backend/src/test/java/com/woowacourse/thankoo/reservation/ReservationControllerTest.java
@@ -1,54 +1,0 @@
-package com.woowacourse.thankoo.reservation;
-
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyLong;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.BDDMockito.given;
-import static org.mockito.Mockito.doNothing;
-import static org.springframework.restdocs.headers.HeaderDocumentation.headerWithName;
-import static org.springframework.restdocs.headers.HeaderDocumentation.requestHeaders;
-import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
-import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.put;
-import static org.springframework.restdocs.payload.JsonFieldType.STRING;
-import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
-import static org.springframework.restdocs.payload.PayloadDocumentation.requestFields;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-import com.woowacourse.thankoo.common.ControllerTest;
-import com.woowacourse.thankoo.reservation.application.dto.ReservationStatusRequest;
-import org.apache.http.HttpHeaders;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.ResultActions;
-
-@DisplayName("ReservationController 는 ")
-class ReservationControllerTest extends ControllerTest {
-
-    @DisplayName("예약 요청 상태를 변경한다.")
-    @Test
-    void updateReservationStatus() throws Exception {
-        given(jwtTokenProvider.getPayload(anyString()))
-                .willReturn("1");
-        doNothing().when(reservationService).updateStatus(anyLong(), anyLong(), any(ReservationStatusRequest.class));
-
-        ResultActions resultActions = mockMvc.perform(put("/api/reservations/1")
-                        .header(HttpHeaders.AUTHORIZATION, "Bearer accessToken")
-                        .accept(MediaType.APPLICATION_JSON)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsString(new ReservationStatusRequest("accept"))))
-                .andDo(print())
-                .andExpect(
-                        status().isNoContent());
-
-        resultActions.andDo(document("reservations/update-reservation-status",
-                getResponsePreprocessor(),
-                requestHeaders(
-                        headerWithName(HttpHeaders.AUTHORIZATION).description("token")
-                ),
-                requestFields(
-                        fieldWithPath("status").type(STRING).description("status")
-                )));
-    }
-}
