+++ backend/src/test/java/com/woowacourse/thankoo/reservation/presentation/ReservationControllerTest.java
@@ -1,11 +1,15 @@
 package com.woowacourse.thankoo.reservation.presentation;
 
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.doNothing;
 import static org.springframework.restdocs.headers.HeaderDocumentation.headerWithName;
 import static org.springframework.restdocs.headers.HeaderDocumentation.requestHeaders;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.post;
+import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.put;
 import static org.springframework.restdocs.payload.JsonFieldType.NUMBER;
 import static org.springframework.restdocs.payload.JsonFieldType.STRING;
 import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
@@ -15,6 +19,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import com.woowacourse.thankoo.common.ControllerTest;
 import com.woowacourse.thankoo.reservation.application.dto.ReservationRequest;
+import com.woowacourse.thankoo.reservation.application.dto.ReservationStatusRequest;
 import java.time.LocalDateTime;
 import org.apache.http.HttpHeaders;
 import org.junit.jupiter.api.DisplayName;
@@ -38,7 +43,7 @@ class ReservationControllerTest extends ControllerTest {
                 .andDo(print())
                 .andExpect(status().isCreated());
 
-        resultActions.andDo(document("reservation/reserve",
+        resultActions.andDo(document("reservations/reserve",
                 getRequestPreprocessor(),
                 requestHeaders(
                         headerWithName(HttpHeaders.AUTHORIZATION).description("token")
@@ -50,4 +55,29 @@ class ReservationControllerTest extends ControllerTest {
         ));
     }
 
+    @DisplayName("예약 요청 상태를 변경한다.")
+    @Test
+    void updateReservationStatus() throws Exception {
+        given(jwtTokenProvider.getPayload(anyString()))
+                .willReturn("1");
+        doNothing().when(reservationService).updateStatus(anyLong(), anyLong(), any(ReservationStatusRequest.class));
+
+        ResultActions resultActions = mockMvc.perform(put("/api/reservations/1")
+                        .header(HttpHeaders.AUTHORIZATION, "Bearer accessToken")
+                        .accept(MediaType.APPLICATION_JSON)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(new ReservationStatusRequest("accept"))))
+                .andDo(print())
+                .andExpect(
+                        status().isNoContent());
+
+        resultActions.andDo(document("reservations/update-reservation-status",
+                getResponsePreprocessor(),
+                requestHeaders(
+                        headerWithName(HttpHeaders.AUTHORIZATION).description("token")
+                ),
+                requestFields(
+                        fieldWithPath("status").type(STRING).description("status")
+                )));
+    }
 }
